backend_mock_solution/requirements.txt
PyMuPDF
Pillow
opencv-python
numpy